+++ R/expM.R
@@ -1,0 +1,8 @@
+expM <- function(X, 
+                 e) {
+    
+    v	=  La.svd(X)  
+    res	=  v$u %*% diag(v$d^e) %*% v$vt
+    
+    return(res)
+}